+++ src/TerminalSandbox.mli
@@ -1,0 +1,7 @@
+type t
+
+val create : Toolchain.resources -> t option
+
+val dispose : t -> unit
+
+val show : t -> unit