+++ src/Toolchain.ml
@@ -109,7 +109,7 @@ end
 
 type resources = PackageManager.t
 
-let toString = PackageManager.toString
+let toString t = "OCaml Platform | " ^ PackageManager.toString t
 
 let availablePackageManagers () =
   { PackageManager.Kind.Hmap.opam = Opam.make ()
@@ -132,7 +132,8 @@ let ofSettings () : PackageManager.t option Promise.t =
       this_ this_
   in
   match
-    (Settings.get PackageManager.Setting.t : PackageManager.Setting.t option)
+    ( Settings.get ~section:"ocaml" PackageManager.Setting.t
+      : PackageManager.Setting.t option )
   with
   | None -> Promise.return None
   | Some (Esy manifest) -> (
@@ -160,7 +161,8 @@ let ofSettings () : PackageManager.t option Promise.t =
     Promise.return (Some PackageManager.Global)
 
 let toSettings (pm : PackageManager.t) =
-  Settings.set PackageManager.Setting.t (PackageManager.toSetting pm)
+  Settings.set ~section:"ocaml" PackageManager.Setting.t
+    (PackageManager.toSetting pm)
 
 module Candidate = struct
   type t =
@@ -247,44 +249,42 @@ let setupToolChain (kind : PackageManager.t) =
 
 let makeResources kind = kind
 
-let selectAndSave () =
+let select () =
   let open Promise.O in
   let workspaceFolders = Vscode.Workspace.workspaceFolders () in
   sandboxCandidates ~workspaceFolders >>= fun candidates ->
-  selectPackageManager candidates >>| function
-  | None -> None
-  | Some choice -> (
-    match choice.status with
-    | Error s ->
-      message `Warn "This toolchain is invalid. Error: %s" s;
-      None
-    | Ok () ->
-      let (_ : unit Promise.t) = toSettings choice.packageManager in
-      Some choice.packageManager )
+  let open Promise.Option.O in
+  selectPackageManager candidates >>= fun { status; packageManager } ->
+  match status with
+  | Error s ->
+    message `Warn "This toolchain is invalid. Error: %s" s;
+    Promise.return None
+  | Ok () -> Promise.Option.return packageManager
 
-let getLspCommand (t : PackageManager.t) : Path.t * string array =
-  let name = "ocamllsp" in
-  match t with
-  | Opam (opam, switch) -> Opam.exec opam ~switch ~args:[| name |]
-  | Esy (esy, manifest) -> Esy.exec esy ~manifest ~args:[| name |]
-  | Global -> (Path.ofString name, [||])
-
-let addLspCheckArg args = Array.append args [| "--help=plain" |]
-
-let getDuneFormatter (t : PackageManager.t) : Path.t * string array =
-  let dune = "dune" in
-  let format = "format-dune-file" in
-  match t with
-  | Opam (opam, switch) -> Opam.exec opam ~switch ~args:[| dune; format |]
-  | Esy (esy, manifest) -> Esy.exec esy ~manifest ~args:[| dune; format |]
-  | Global -> (Path.ofString dune, [| format |])
+let selectAndSave () =
+  let open Promise.Option.O in
+  select () >>= fun packageManager ->
+  let open Promise.O in
+  toSettings packageManager >>| fun () -> Some packageManager
 
 let getCommand (t : PackageManager.t) bin args : Path.t * string array =
   let binArgs = Array.of_list (bin :: args) in
-  match t with
-  | Opam (opam, switch) -> Opam.exec opam ~switch ~args:binArgs
-  | Esy (esy, manifest) -> Esy.exec esy ~manifest ~args:binArgs
-  | Global -> (Path.ofString bin, Array.of_list args)
+  let bin, args =
+    match t with
+    | Opam (opam, switch) -> Opam.exec opam ~switch ~args:binArgs
+    | Esy (esy, manifest) -> Esy.exec esy ~manifest ~args:binArgs
+    | Global -> (Path.ofString bin, Array.of_list args)
+  in
+  log "getCommand: %s %s" (Path.toString bin) (Js.Array.joinWith " " args);
+  (bin, args)
+
+let getLspCommand (t : PackageManager.t) : Path.t * string array =
+  getCommand t "ocamllsp" []
+
+let getDuneCommand (t : PackageManager.t) args : Path.t * string array =
+  getCommand t "dune" args
+
+let addLspCheckArg args = Array.append args [| "--help=plain" |]
 
 let runSetup resources =
   let open Promise.Result.O in