+++ src/Toolchain.mli
@@ -37,8 +37,10 @@ val makeResources : PackageManager.t -> resources
    (ie VSCode's process environment) and the project root and produces a promise
    of resources available that can later be passed on to runSetup that can be
    called to install the toolchain. *)
-
 val selectAndSave : unit -> PackageManager.t option Promise.t
+
+(** [select] is the same as [selectAndSave] but does not save the toolchain configuration *)
+val select : unit -> PackageManager.t option Promise.t
 
 (** [runSetup] is a effectful function that triggers setup instructions
    automatically for the user. At present, this functionality
@@ -70,12 +72,12 @@ val runSetup : resources -> (unit, string) result Promise.t
 
 (* Helper utils *)
 
+(** Extract command to run with the toolchain *)
+val getCommand : resources -> string -> string list -> Path.t * string array
+
 (** Extract lsp command and arguments (Eg. "opam" and [| "exec";
    "ocamllsp" |] *)
 val getLspCommand : resources -> Path.t * string array
 
-(** Extract format-dune-file command *)
-val getDuneFormatter : resources -> Path.t * string array
-
-(** Extract command to run with the toolchain *)
-val getCommand : resources -> string -> string list -> Path.t * string array
+(** Extract a dune command *)
+val getDuneCommand : resources -> string list -> Path.t * string array