+++ src/bindings/Vscode.mli
@@ -174,7 +174,7 @@ module Window : sig
 
   and rangeEdge = { character : int }
 
-  val activeTextEditor : activeTextEditor option
+  val activeTextEditor : unit -> activeTextEditor option
 
   type location =
     | SourceControl
@@ -211,6 +211,25 @@ module Window : sig
        ?options:textDocumentShowOptions
     -> [ `Uri of TextDocument.uri | `Document of TextDocument.t ]
     -> TextEditor.t Promise.t
+
+  module Terminal : sig
+    type t
+
+    val dispose : t -> unit
+
+    val hide : t -> unit
+
+    val sendText : t -> string -> ?addNewLine:bool -> unit -> unit
+
+    val show : t -> ?preserveFocus:bool -> unit -> unit
+  end
+
+  val createTerminal :
+       ?name:string
+    -> ?shellPath:string
+    -> ?shellArgs:string array
+    -> unit
+    -> Terminal.t
 end
 
 module Folder : sig
@@ -234,7 +253,7 @@ module Workspace : sig
 
   type cancellationToken = { isCancellationRequested : bool }
 
-  val name : string option
+  val name : unit -> string option
 
   val workspaceFolders : unit -> Folder.t array
 
@@ -245,9 +264,9 @@ module Workspace : sig
   val onDidChangeWorkspaceFolders :
     (workspaceFoldersChangeEvent -> unit) -> unit
 
-  val textDocuments : TextDocument.event array
+  val textDocuments : unit -> TextDocument.event array
 
-  val getConfiguration : string -> WorkspaceConfiguration.t
+  val getConfiguration : ?section:string -> unit -> WorkspaceConfiguration.t
 
   val findFiles :
        inc:string
@@ -448,4 +467,8 @@ end
 module Tasks : sig
   val registerTaskProvider :
     typ:string -> provider:TaskProvider.t -> Disposable.t
+end
+
+module Env : sig
+  val shell : unit -> string
 end