+++ src/settings.mli
@@ -2,9 +2,9 @@ open Import
 
 type 'a t
 
-val get : 'a t -> 'a option
+val get : ?section:string -> 'a t -> 'a option
 
-val set : 'a t -> 'a -> unit Promise.t
+val set : ?section:string -> 'a t -> 'a -> unit Promise.t
 
 val create :
      scope:WorkspaceConfiguration.configurationTarget